New example and bugfixes
09_MultipleLights: I added a new example to demonstrate multiple ligts with a per-fragment-shader program.
Note that this replaces the former example, which can be still found in 09_MultipleLights_depracted. But it will disappear in future.

Model___.h/.cpp: All model classes have a new api getProgam() to return the shader program idx.

CommonTypes.h: Contains datatype delcarations for materials and lights. It replaces the deprecated Appearance class.